// web/nestspend-web/src/app/core/api/models/transaction-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Transaction response data
 */
export interface TransactionResponse {

  /**
   * UUID of the account
   */
  accountId?: string;

  /**
   * Name of the account
   */
  accountName?: string;

  /**
   * Amount in cents
   */
  amountCents?: number;

  /**
   * UUID of the category
   */
  categoryId?: string;

  /**
   * Name of the category
   */
  categoryName?: string;

  /**
   * Timestamp when the transaction was created
   */
  createdAt?: string;

  /**
   * UUID of the user who created the transaction
   */
  createdById?: string;

  /**
   * Display name of the user who created the transaction
   */
  createdByName?: string;

  /**
   * Unique identifier of the transaction
   */
  id?: string;

  /**
   * Merchant name
   */
  merchant?: string;

  /**
   * Additional note
   */
  note?: string;

  /**
   * Date of the transaction
   */
  txDate?: string;

  /**
   * Type of transaction
   */
  type?: 'EXPENSE' | 'INCOME';

  /**
   * Timestamp when the transaction was last updated
   */
  updatedAt?: string;
}