// web/nestspend-web/src/app/core/api/models/account-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Account response data
 */
export interface AccountResponse {

  /**
   * Timestamp when the account was created
   */
  createdAt?: string;

  /**
   * Unique identifier of the account
   */
  id?: string;

  /**
   * Name of the account
   */
  name?: string;

  /**
   * Type of account
   */
  type?: 'CASH' | 'BANK' | 'CARD';
}